+++ QEfficient/base/onnx_transforms.py
@@ -91,49 +91,3 @@ class SplitTensorsTransform(OnnxTransform):
                     current_file_size = tsize
                 external_data_helper.set_external_data(tensor, f"{model_name}_{file_num}.onnx.data")
         return model, transformed
-
-
-class RemoveCrossAttentionIOTransform(OnnxTransform):
-    """
-    Removes the input and output names of cross-attention layers.
-    """
-
-    @classmethod
-    def apply(cls, model: ModelProto, *, onnx_base_dir: Optional[str] = None, **kwargs) -> Tuple[ModelProto, bool]:
-        """
-        :param onnx_base_dir: Base directory to load tensors (if not already loaded).
-        """
-        layers_to_remove = [3, 8, 13, 18, 23, 28, 33, 38]
-        names_to_remove = []
-        for layer in layers_to_remove:
-            names_to_remove.append(f"past_key.{layer}_RetainedState")
-            names_to_remove.append(f"past_value.{layer}_RetainedState")
-            names_to_remove.append(f"past_key.{layer}")
-            names_to_remove.append(f"past_value.{layer}")
-
-        graph = model.graph
-        transformed = False
-
-        # Remove outputs
-        for name in names_to_remove:
-            output_to_remove = None
-            for output in graph.output:
-                if output.name == name:
-                    output_to_remove = output
-                    break
-            if output_to_remove:
-                graph.output.remove(output_to_remove)
-                transformed = True
-
-        # # Remove inputs
-        # for name in names_to_remove:
-        #     input_to_remove = None
-        #     for input in graph.input:
-        #         if input.name == name:
-        #             input_to_remove = input
-        #             break
-        #     if input_to_remove:
-        #         graph.input.remove(input_to_remove)
-        #         transformed = True
-
-        return model, transformed
